\name{cjr}
\alias{cjr}
\docType{package}
\title{
Routines for combined judgmental random sampling (CJR)
}
\description{
Collection of functions for calculating Bayesian probabilities and
sample sizes for the CJR methodology
}
\details{
\tabular{ll}{
Package: \tab cjr\cr
Type: \tab Package\cr
Version: \tab 2010.03.18\cr
Date: \tab 2010-03-18\cr
License: \tab PNNL business sensitive\cr
}
}
\author{
Landon Sego
}
% \references{
% ~~ Literature or other references for background information ~~
% }
\keyword{package}
% ~~ Optionally other keywords from doc/KEYWORDS, one per line
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
