% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{computeHotspot}
\alias{computeHotspot}
\title{Run Hotspot Analysis}
\usage{
computeHotspot(expression, neighbors, weights)
}
\arguments{
\item{expression}{Matrix on which to calculate G_i. 
G_i is calculated for each variable (row) separately.  matrix of dimension 
N_GENES x N_CELLS}

\item{neighbors}{n_nearest neighbors for each cell.  Matrix of
size N_CELLS x N_Neighbors.  Entries represent index of neighbors}

\item{weights}{weights for each neighbor.  Same size/type as neighbors
input}
}
\value{
G_i Getis-ord values for each variable in G_i.  Matrix of size
N_GENES x N_CELLS
}
\description{
Compute the Getis-Ord coefficient for variables in expression using
the provided connectivity (neighbors) and associated edge weights
}
