% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{computeWeights}
\alias{computeWeights}
\title{Compute weights for neighbors}
\usage{
computeWeights(distances, neighborhood_factor = 3)
}
\arguments{
\item{distances}{distances to neighbors.  matrix of dimension
N_CELLS x N_NEIGHBORS}

\item{neighborhood_factor}{What proportion of neighborhood to use to
define the gaussian kernel width. Default is 3}
}
\value{
weights weights for each neighbor.  Same size/type as distances
input
}
\description{
Given a matrix of distances, computes appropriate 'weights'
which decay according to a Guassian Kernel with width equal
to the distance to the N_NEIGHBORS / <neighborhood_factor> neighbor
}
