\name{hotspot-package}
\alias{hotspot-package}
\alias{hotspot}
\docType{package}
\title{
\packageTitle{hotspot}
}
\description{
\packageDescription{hotspot}
}
\details{

The package contains results of hotspot analysis of some open source system, and a function which performs the test itself.
See \link{hotspot.data} for details about the input data, and \link{perform.hotspot.tests} for the execution of tests.
}
\author{
\packageMaintainer{hotspot}
}
