% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{neighborsAndWeights}
\alias{neighborsAndWeights}
\title{Find Nearest Neighbors and Asociated Weights}
\usage{
neighborsAndWeights(data, n_neighbors = 30, neighborhood_factor = 3)
}
\arguments{
\item{data}{Matrix to use to derive distances. It's recommended to u
se a PCA-reduced gene expression matrix here.  matrix of dimension 
N_CELLS x N_COMPONENTS}

\item{n_neighbors}{How many neighbors to use for each cell}

\item{neighborhood_factor}{What proportion of neighborhood to use to
define the gaussian kernel width. Default is 3}
}
\value{
out$neighbors n_nearest neighbors for each cell.  Matrix of
size N_CELLS x N_Neighbors.  Entries represent index of neighbors

out$weights weights for each neighbor.  Same size/type as out$neighbors
input
}
\description{
Using the input data matrix, for each cell (row) compute the
nearest n_neighbors and compute a weight for each neighbor.
}
